refactor inverted index
